+++ test/eca/llm_api_test.clj
@@ -26,13 +26,13 @@
     (with-redefs [config/get-env (constantly nil)]
       (let [db {:models {}}
             config {:providers {"anthropic" {:key "something"}}}]
-        (is (= "anthropic/claude-sonnet-4-20250514" (llm-api/default-model db config))))))
+        (is (= "anthropic/claude-sonnet-4-5-20250929" (llm-api/default-model db config))))))
 
   (testing "Anthropic API key present in ENV"
     (with-redefs [config/get-env (fn [k] (when (= k "ANTHROPIC_API_KEY") "env-anthropic"))]
       (let [db {:models {}}
             config {:providers {"anthropic" {:keyEnv "ANTHROPIC_API_KEY"}}}]
-        (is (= "anthropic/claude-sonnet-4-20250514" (llm-api/default-model db config))))))
+        (is (= "anthropic/claude-sonnet-4-5-20250929" (llm-api/default-model db config))))))
 
   (testing "OpenAI API key present in config"
     (with-redefs [config/get-env (constantly nil)]
@@ -50,4 +50,4 @@
     (with-redefs [config/get-env (constantly nil)]
       (let [db {:models {}}
             config {}]
-        (is (= "anthropic/claude-sonnet-4-20250514" (llm-api/default-model db config)))))))
+        (is (= "anthropic/claude-sonnet-4-5-20250929" (llm-api/default-model db config)))))))